+++ victreebot/extentions/events/member_create_add.py
@@ -8,18 +8,23 @@
 # IMPORTS
 import os
 
+import hikari
 import tanjun
+from core.bot import Bot
 from dotenv import load_dotenv
-from extentions.interactions.profile.edit import command_profile_edit
-from extentions.interactions.profile.view import command_profile_view
+from utils.DatabaseHandler import DatabaseHandler
+from utils.helpers.BotUtils import BotUtils
 
 load_dotenv()
 BOT_NAME = os.getenv("BOT_NAME")
 
 
-# ------------------------------------------------------------------------- #
-# THE BOTS SUB COMMANDS GROUP (E.G. /{sub_command_group}) #
-# ------------------------------------------------------------------------- #
-profile_group = (
-    tanjun.slash_command_group("profile", ".").add_command(command_profile_edit).add_command(command_profile_view)
-)
+async def event_member_create_add(
+    event: hikari.MemberCreateEvent,
+    db: DatabaseHandler = tanjun.injected(type=DatabaseHandler),
+    bot: BotUtils = tanjun.injected(type=BotUtils),
+    bot_aware: Bot = tanjun.injected(type=Bot),
+):
+    guild = await event.app.rest.fetch_guild(event.guild_id)
+    # ADD USER
+    await db.insert_user(guild, event.user)