+++ victreebot/extentions/events/member_delete_remove.py
@@ -8,18 +8,22 @@
 # IMPORTS
 import os
 
+import hikari
 import tanjun
+from core.bot import Bot
 from dotenv import load_dotenv
-from extentions.interactions.profile.edit import command_profile_edit
-from extentions.interactions.profile.view import command_profile_view
+from utils.DatabaseHandler import DatabaseHandler
+from utils.helpers.BotUtils import BotUtils
 
 load_dotenv()
 BOT_NAME = os.getenv("BOT_NAME")
 
 
-# ------------------------------------------------------------------------- #
-# THE BOTS SUB COMMANDS GROUP (E.G. /{sub_command_group}) #
-# ------------------------------------------------------------------------- #
-profile_group = (
-    tanjun.slash_command_group("profile", ".").add_command(command_profile_edit).add_command(command_profile_view)
-)
+async def event_member_delete_remove(
+    event: hikari.MemberCreateEvent,
+    db: DatabaseHandler = tanjun.injected(type=DatabaseHandler),
+    bot: BotUtils = tanjun.injected(type=BotUtils),
+    bot_aware: Bot = tanjun.injected(type=Bot),
+):
+    # REMOVE USER
+    await db.delete_user(event.guild_id, event.user_id)