+++ victreebot/extentions/interactions/profile/edit.py
@@ -1,10 +1,9 @@
 # ------------------------------------------------------------------------- #
-# Husqy Discord Bot, see https://www.husqy.xyz for more info                #
+# VictreeBot                                                                #
 #                                                                           #
-# Copyright (C) Husqy - All Rights Reserved                                 #
-# Unauthorized copying of this file, via any medium is strictly prohibited  #
-# Proprietary and confidential                                              #
-# Written by Luke Hendriks <luke@la-online.nl>, July 2021 (C)               #
+# See LICENSE for more information. If this code is used, attribution       #
+# would be appreciated.                                                     #
+# Written by Luke Hendriks                                                  #
 # ------------------------------------------------------------------------- #
 # IMPORTS
 import asyncio
@@ -12,11 +11,11 @@ import os
 
 import hikari
 import tanjun
+from core.bot import Bot
 from dotenv import load_dotenv
 from utils.DatabaseHandler import DatabaseHandler
 from utils.helpers.BotUtils import BotUtils
 from utils.helpers.contants import SUPPORTED_LANGUAGES
-from core.bot import Bot
 
 load_dotenv()
 BOT_NAME = os.getenv("BOT_NAME")
@@ -31,7 +30,6 @@ async def command_profile_edit(
     db: DatabaseHandler = tanjun.injected(type=DatabaseHandler),
     bot: BotUtils = tanjun.injected(type=BotUtils),
     bot_aware: Bot = tanjun.injected(type=Bot),
-
 ):
     language, auto_delete, gmt, *none = await db.get_guild_settings(
         guild=ctx.get_guild(), settings=["language", "auto_delete", "gmt"]
@@ -41,11 +39,9 @@ async def command_profile_edit(
     )
 
     timeout = 120
-    embed = (
-        hikari.Embed(
-            title=SUPPORTED_LANGUAGES.get(language).profile_edit_embed_title,
-            description=SUPPORTED_LANGUAGES.get(language).profile_edit_embed_description,
-        )
+    embed = hikari.Embed(
+        title=SUPPORTED_LANGUAGES.get(language).profile_edit_embed_title,
+        description=SUPPORTED_LANGUAGES.get(language).profile_edit_embed_description,
     )
 
     action_row_1 = (
@@ -73,14 +69,17 @@ async def command_profile_edit(
         event = await ctx.client.events.wait_for(
             hikari.InteractionCreateEvent,
             timeout=timeout,
-            predicate=lambda event: event.interaction.user.id == ctx.author.id
+            predicate=lambda event: isinstance(event.interaction, hikari.ComponentInteraction)
+            and event.interaction.user.id == ctx.author.id
             and event.interaction.message.id == response_message.id,
         )
     except asyncio.TimeoutError:
         response = SUPPORTED_LANGUAGES.get(language).response_profile_edit_timeout_reached
         await ctx.edit_last_response(response, delete_after=auto_delete, embed=None, components=None)
         if log_errors:
-            log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit_timeout_reached.format(datetime=await bot.get_timestamp_aware(gmt), member=ctx.member)
+            log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit_timeout_reached.format(
+                datetime=await bot.get_timestamp_aware(gmt), member=ctx.member
+            )
             await bot.log_from_ctx(ctx, db, log_response)
         return
     else:
@@ -91,7 +90,9 @@ async def command_profile_edit(
                     label=SUPPORTED_LANGUAGES.get(language).profile_modal_add_friend_codes_text_input_title,
                     custom_id="friend_codes",
                 )
-                .set_placeholder(SUPPORTED_LANGUAGES.get(language).profile_modal_add_friend_codes_text_input_placeholder)
+                .set_placeholder(
+                    SUPPORTED_LANGUAGES.get(language).profile_modal_add_friend_codes_text_input_placeholder
+                )
                 .add_to_container()
             )
             await event.interaction.create_modal_response(
@@ -104,33 +105,42 @@ async def command_profile_edit(
                 event = await ctx.client.events.wait_for(
                     hikari.InteractionCreateEvent,
                     timeout=timeout,
-                    predicate=lambda event: event.interaction.user.id == ctx.author.id,
+                    predicate=lambda event: isinstance(event.interaction, hikari.ModalInteraction)
+                    and event.interaction.user.id == ctx.author.id,
                 )
             except asyncio.TimeoutError:
                 response = SUPPORTED_LANGUAGES.get(language).response_profile_edit_timeout_reached
                 await ctx.edit_last_response(response, delete_after=auto_delete, embed=None, components=None)
                 if log_errors:
-                    log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit_timeout_reached.format(datetime=await bot.get_timestamp_aware(gmt), member=ctx.member)
+                    log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit_timeout_reached.format(
+                        datetime=await bot.get_timestamp_aware(gmt), member=ctx.member
+                    )
                     await bot.log_from_ctx(ctx, db, log_response)
                 return
             else:
                 await event.interaction.create_initial_response(6)
                 if event.interaction.custom_id == "friend_codes_modal":
                     if event.interaction.components[0][0].value is not None:
-                        current_friend_codes, *none = await db.get_user_details(ctx.get_guild(), ctx.member, details=["friend_codes"])
+                        current_friend_codes, *none = await db.get_user_details(
+                            ctx.get_guild(), ctx.member, details=["friend_codes"]
+                        )
                         if current_friend_codes is not None:
                             current_friend_codes = current_friend_codes.strip("'")
                             current_friend_codes_list = current_friend_codes.split(",")
 
-                        if current_friend_codes is None or current_friend_codes_list is None or current_friend_codes_list == []:
+                        if (
+                            current_friend_codes is None
+                            or current_friend_codes_list is None
+                            or current_friend_codes_list == []
+                        ):
                             current_friend_codes_list = []
-                            
+
                         friend_codes_list = event.interaction.components[0][0].value.split(",")
                         for friend_code in friend_codes_list:
-                            friend_code = friend_code.replace("-"," ")
+                            friend_code = friend_code.replace("-", " ")
                             friend_code = friend_code.lstrip()
                             friend_code = friend_code.rstrip()
-                            if friend_code not in current_friend_codes_list:   
+                            if friend_code not in current_friend_codes_list:
                                 current_friend_codes_list.append(friend_code)
 
                         friend_codes_to_add = f'{",".join(friend_code for friend_code in current_friend_codes_list)}'
@@ -142,9 +152,11 @@ async def command_profile_edit(
                         response = SUPPORTED_LANGUAGES.get(language).response_profile_edit_success
                         await ctx.edit_last_response(response, delete_after=auto_delete, embed=None, components=None)
                         if log_profile_edit:
-                            log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit.format(datetime=await bot.get_timestamp_aware(gmt), member=ctx.member)
+                            log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit.format(
+                                datetime=await bot.get_timestamp_aware(gmt), member=ctx.member
+                            )
                             await bot.log_from_ctx(ctx, db, log_response)
-                            
+
         elif event.interaction.custom_id == "delete_friend_codes":
             delete_friend_code_action_row = (
                 ctx.rest.build_action_row()
@@ -152,7 +164,9 @@ async def command_profile_edit(
                     label=SUPPORTED_LANGUAGES.get(language).profile_modal_delete_friend_codes_text_input_title,
                     custom_id="friend_codes",
                 )
-                .set_placeholder(SUPPORTED_LANGUAGES.get(language).profile_modal_delete_friend_codes_text_input_placeholder)
+                .set_placeholder(
+                    SUPPORTED_LANGUAGES.get(language).profile_modal_delete_friend_codes_text_input_placeholder
+                )
                 .add_to_container()
             )
             await event.interaction.create_modal_response(
@@ -165,40 +179,51 @@ async def command_profile_edit(
                 event = await ctx.client.events.wait_for(
                     hikari.InteractionCreateEvent,
                     timeout=timeout,
-                    predicate=lambda event: event.interaction.user.id == ctx.author.id,
+                    predicate=lambda event: isinstance(event.interaction, hikari.ModalInteraction)
+                    and event.interaction.user.id == ctx.author.id,
                 )
             except asyncio.TimeoutError:
                 response = SUPPORTED_LANGUAGES.get(language).response_profile_edit_timeout_reached
                 await ctx.edit_last_response(response, delete_after=auto_delete, embed=None, components=None)
                 if log_errors:
-                    log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit_timeout_reached.format(datetime=await bot.get_timestamp_aware(gmt), member=ctx.member)
+                    log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit_timeout_reached.format(
+                        datetime=await bot.get_timestamp_aware(gmt), member=ctx.member
+                    )
                     await bot.log_from_ctx(ctx, db, log_response)
                 return
             else:
                 await event.interaction.create_initial_response(6)
                 if event.interaction.custom_id == "friend_codes_modal":
                     if event.interaction.components[0][0].value is not None:
-                        current_friend_codes, *none = await db.get_user_details(ctx.get_guild(), ctx.member, details=["friend_codes"])
+                        current_friend_codes, *none = await db.get_user_details(
+                            ctx.get_guild(), ctx.member, details=["friend_codes"]
+                        )
                         if current_friend_codes is not None:
                             current_friend_codes = current_friend_codes.strip("'")
                             current_friend_codes_list = current_friend_codes.split(",")
 
-                        if current_friend_codes is None or current_friend_codes_list is None or current_friend_codes_list == []:
+                        if (
+                            current_friend_codes is None
+                            or current_friend_codes_list is None
+                            or current_friend_codes_list == []
+                        ):
                             current_friend_codes_list = []
-                            
+
                         friend_codes_list = event.interaction.components[0][0].value.split(",")
                         for friend_code in friend_codes_list:
-                            friend_code = friend_code.replace("-"," ")
+                            friend_code = friend_code.replace("-", " ")
                             friend_code = friend_code.lstrip()
                             friend_code = friend_code.rstrip()
                             if friend_code in current_friend_codes_list:
                                 current_friend_codes_list.remove(friend_code)
 
                         if current_friend_codes_list != []:
-                            friend_codes_to_delete = f'{",".join(friend_code for friend_code in current_friend_codes_list)}'
+                            friend_codes_to_delete = (
+                                f'{",".join(friend_code for friend_code in current_friend_codes_list)}'
+                            )
                             friend_codes_to_delete = f"'{friend_codes_to_delete}'"
                         else:
-                            friend_codes_to_delete = f"NULL"
+                            friend_codes_to_delete = "NULL"
                         parameters = []
                         parameters.append(f"friend_codes = {friend_codes_to_delete}")
                         await db.set_user_detail(ctx.get_guild(), ctx.member, parameters=parameters)
@@ -206,7 +231,9 @@ async def command_profile_edit(
                         response = SUPPORTED_LANGUAGES.get(language).response_profile_edit_success
                         await ctx.edit_last_response(response, delete_after=auto_delete, embed=None, components=None)
                         if log_profile_edit:
-                            log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit.format(datetime=await bot.get_timestamp_aware(gmt), member=ctx.member)
+                            log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit.format(
+                                datetime=await bot.get_timestamp_aware(gmt), member=ctx.member
+                            )
                             await bot.log_from_ctx(ctx, db, log_response)
 
         elif event.interaction.custom_id == "add_active_locations":
@@ -216,7 +243,9 @@ async def command_profile_edit(
                     label=SUPPORTED_LANGUAGES.get(language).profile_modal_add_active_locations_text_input_title,
                     custom_id="active_locations",
                 )
-                .set_placeholder(SUPPORTED_LANGUAGES.get(language).profile_modal_add_active_locations_text_input_placeholder)
+                .set_placeholder(
+                    SUPPORTED_LANGUAGES.get(language).profile_modal_add_active_locations_text_input_placeholder
+                )
                 .add_to_container()
             )
             await event.interaction.create_modal_response(
@@ -229,35 +258,44 @@ async def command_profile_edit(
                 event = await ctx.client.events.wait_for(
                     hikari.InteractionCreateEvent,
                     timeout=timeout,
-                    predicate=lambda event: event.interaction.user.id == ctx.author.id,
+                    predicate=lambda event: isinstance(event.interaction, hikari.ModalInteraction)
+                    and event.interaction.user.id == ctx.author.id,
                 )
             except asyncio.TimeoutError:
                 response = SUPPORTED_LANGUAGES.get(language).response_profile_edit_timeout_reached
                 await ctx.edit_last_response(response, delete_after=auto_delete, embed=None, components=None)
                 if log_errors:
-                    log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit_timeout_reached.format(datetime=await bot.get_timestamp_aware(gmt), member=ctx.member)
+                    log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit_timeout_reached.format(
+                        datetime=await bot.get_timestamp_aware(gmt), member=ctx.member
+                    )
                     await bot.log_from_ctx(ctx, db, log_response)
                 return
             else:
                 await event.interaction.create_initial_response(6)
                 if event.interaction.custom_id == "active_locations_modal":
                     if event.interaction.components[0][0].value is not None:
-                        current_active_locations, *none = await db.get_user_details(ctx.get_guild(), ctx.member, details=["active_locations"])
+                        current_active_locations, *none = await db.get_user_details(
+                            ctx.get_guild(), ctx.member, details=["active_locations"]
+                        )
                         if current_active_locations is not None:
                             current_active_locations = current_active_locations.strip("'")
                             current_active_locations_list = current_active_locations.split(",")
 
-                        if current_active_locations is None or current_active_locations_list is None or current_active_locations_list == []:
+                        if (
+                            current_active_locations is None
+                            or current_active_locations_list is None
+                            or current_active_locations_list == []
+                        ):
                             current_active_locations_list = []
-                            
+
                         active_locations_list = event.interaction.components[0][0].value.split(",")
                         for location in active_locations_list:
                             location = location.lstrip()
                             location = location.rstrip()
                             location = location.lower()
-                            if location not in current_active_locations_list:                                
+                            if location not in current_active_locations_list:
                                 current_active_locations_list.append(location)
-                        
+
                         active_locations_to_add = f'{",".join(location for location in current_active_locations_list)}'
                         active_locations_to_add = f"'{active_locations_to_add}'"
                         parameters = []
@@ -267,7 +305,9 @@ async def command_profile_edit(
                         response = SUPPORTED_LANGUAGES.get(language).response_profile_edit_success
                         await ctx.edit_last_response(response, delete_after=auto_delete, embed=None, components=None)
                         if log_profile_edit:
-                            log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit.format(datetime=await bot.get_timestamp_aware(gmt), member=ctx.member)
+                            log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit.format(
+                                datetime=await bot.get_timestamp_aware(gmt), member=ctx.member
+                            )
                             await bot.log_from_ctx(ctx, db, log_response)
 
         elif event.interaction.custom_id == "delete_active_locations":
@@ -277,7 +317,9 @@ async def command_profile_edit(
                     label=SUPPORTED_LANGUAGES.get(language).profile_modal_delete_active_locations_text_input_title,
                     custom_id="active_locations",
                 )
-                .set_placeholder(SUPPORTED_LANGUAGES.get(language).profile_modal_delete_active_locations_text_input_placeholder)
+                .set_placeholder(
+                    SUPPORTED_LANGUAGES.get(language).profile_modal_delete_active_locations_text_input_placeholder
+                )
                 .add_to_container()
             )
             await event.interaction.create_modal_response(
@@ -290,40 +332,51 @@ async def command_profile_edit(
                 event = await ctx.client.events.wait_for(
                     hikari.InteractionCreateEvent,
                     timeout=timeout,
-                    predicate=lambda event: event.interaction.user.id == ctx.author.id,
+                    predicate=lambda event: isinstance(event.interaction, hikari.ModalInteraction)
+                    and event.interaction.user.id == ctx.author.id,
                 )
             except asyncio.TimeoutError:
                 response = SUPPORTED_LANGUAGES.get(language).response_profile_edit_timeout_reached
                 await ctx.edit_last_response(response, delete_after=auto_delete, embed=None, components=None)
                 if log_errors:
-                    log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit_timeout_reached.format(datetime=await bot.get_timestamp_aware(gmt), member=ctx.member)
+                    log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit_timeout_reached.format(
+                        datetime=await bot.get_timestamp_aware(gmt), member=ctx.member
+                    )
                     await bot.log_from_ctx(ctx, db, log_response)
                 return
             else:
                 await event.interaction.create_initial_response(6)
                 if event.interaction.custom_id == "active_locations_modal":
                     if event.interaction.components[0][0].value is not None:
-                        current_active_locations, *none = await db.get_user_details(ctx.get_guild(), ctx.member, details=["active_locations"])
+                        current_active_locations, *none = await db.get_user_details(
+                            ctx.get_guild(), ctx.member, details=["active_locations"]
+                        )
                         if current_active_locations is not None:
                             current_active_locations = current_active_locations.strip("'")
                             current_active_locations_list = current_active_locations.split(",")
 
-                        if current_active_locations is None or current_active_locations_list is None or current_active_locations_list == []:
+                        if (
+                            current_active_locations is None
+                            or current_active_locations_list is None
+                            or current_active_locations_list == []
+                        ):
                             current_active_locations_list = []
-                            
+
                         active_locations_list = event.interaction.components[0][0].value.split(",")
                         for location in active_locations_list:
                             location = location.lstrip()
                             location = location.rstrip()
                             location = location.lower()
-                            if location in current_active_locations_list:                                
+                            if location in current_active_locations_list:
                                 current_active_locations_list.remove(location)
-                        
+
                         if current_active_locations_list != []:
-                            active_locations_to_delete = f'{",".join(location for location in current_active_locations_list)}'
+                            active_locations_to_delete = (
+                                f'{",".join(location for location in current_active_locations_list)}'
+                            )
                             active_locations_to_delete = f"'{active_locations_to_delete}'"
                         else:
-                            active_locations_to_delete = f"NULL"
+                            active_locations_to_delete = "NULL"
 
                         parameters = []
                         parameters.append(f"active_locations = {active_locations_to_delete}")
@@ -332,5 +385,7 @@ async def command_profile_edit(
                         response = SUPPORTED_LANGUAGES.get(language).response_profile_edit_success
                         await ctx.edit_last_response(response, delete_after=auto_delete, embed=None, components=None)
                         if log_profile_edit:
-                            log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit.format(datetime=await bot.get_timestamp_aware(gmt), member=ctx.member)
+                            log_response = SUPPORTED_LANGUAGES.get(language).log_response_profile_edit.format(
+                                datetime=await bot.get_timestamp_aware(gmt), member=ctx.member
+                            )
                             await bot.log_from_ctx(ctx, db, log_response)